fix: move image and JS to fixtures dir
