+++ packages/components/bolt-figure/__tests__/figure.js
@@ -1,5 +1,5 @@
 const { render } = require('@bolt/twig-renderer');
-const { media } = require('./figure-data');
+const { media } = require('./fixtures/figure-data');
 
 describe('figure', async () => {
   Object.keys(media).forEach(async item => {
